test: 팀의 스케줄을 가져오는 테스트 코드 위치 변경
