+++ src/test/kotlin/com/fcdebug/earlysoccerserver/schedule/ScheduleTest.kt
@@ -5,6 +5,7 @@ import com.fcdebug.earlysoccerserver.domain.schedule.ScheduleRepository
 import com.fcdebug.earlysoccerserver.domain.team.Team
 import com.fcdebug.earlysoccerserver.domain.team.TeamRepository
 import io.github.serpro69.kfaker.faker
+import jakarta.transaction.Transactional
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import org.springframework.test.context.ActiveProfiles
@@ -19,8 +20,34 @@ class ScheduleTest @Autowired constructor (
     private val scheduleRepository: ScheduleRepository,
 ) {
     val faker = faker {  }
+
     @Test
-    fun `스케줄을 생성한다`() {
+    @Transactional
+    fun `팀의 전체 스케줄을 가져온다`() {
+        //given
+        val team: Team = teamRepository.save(
+            Team.create(faker.team.name())
+        )
+
+        //when
+        val schedules = listOf(
+            Schedule.create(
+                team = team,
+                date = LocalDateTime.now(),
+                place = faker.address.fullAddress(),
+                opponent = faker.team.name(),))
+        scheduleRepository.saveAll(schedules)
+        val saveSchedules: List<Schedule> = scheduleRepository.findByTeamId(team.toDto().id!!)
+
+        //then
+        assertThat(saveSchedules[0].team.name).isEqualTo(schedules[0].team.name)
+        assertThat(saveSchedules[0].date).isEqualTo(schedules[0].date)
+        assertThat(saveSchedules[0].place).isEqualTo(schedules[0].place)
+        assertThat(saveSchedules[0].opponent).isEqualTo(schedules[0].opponent)
+    }
+
+    @Test
+    fun `팀의 스케줄을 생성한다`() {
         //given
         val team: Team = teamRepository.save(
             Team.create(faker.team.name())
