+++ src/test/kotlin/com/fcdebug/earlysoccerserver/team/TeamTest.kt
@@ -11,39 +11,16 @@ import jakarta.transaction.Transactional
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.ActiveProfiles
 import java.time.LocalDateTime
 
 @ActiveProfiles("test")
-@SpringBootTest
+@DataJpaTest
 class TeamTest @Autowired constructor (
     private val teamRepository: TeamRepository,
-    private val scheduleRepository: ScheduleRepository,
 ) {
     val faker = faker {  }
-    @Test
-    @Transactional
-    fun `팀의 전체 스케줄을 가져온다`() {
-        //given
-        val team: Team = teamRepository.save(
-            Team.create(faker.team.name())
-        )
 
-        //when
-        val schedules = listOf(
-            Schedule.create(
-                team = team,
-                date = LocalDateTime.now(),
-                place = faker.address.fullAddress(),
-                opponent = faker.team.name(),))
-        scheduleRepository.saveAll(schedules)
-        val saveSchedules: List<Schedule> = scheduleRepository.findByTeamId(team.toDto().id!!)
-
-        //then
-        assertThat(saveSchedules[0].team.name).isEqualTo(schedules[0].team.name)
-        assertThat(saveSchedules[0].date).isEqualTo(schedules[0].date)
-        assertThat(saveSchedules[0].place).isEqualTo(schedules[0].place)
-        assertThat(saveSchedules[0].opponent).isEqualTo(schedules[0].opponent)
-    }
 }
